pyproject.toml
[project]
name = "claude-agentcore-workshop"
version = "0.1.0"
description = "Workshop: Build Agentic AI Applications with Claude Agent SDK and Amazon Bedrock AgentCore"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "AWS Workshop"}
]
keywords = ["ai", "agent", "claude", "bedrock", "agentcore", "workshop"]

dependencies = [
    "claude-agent-sdk>=0.1.19",
    "anyio>=3.7.0",
    "python-dotenv>=1.0.0",
    "bedrock-agentcore>=1.1.4",
    "boto3>=1.34.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "faker>=24.0.0",
    "sqlparse>=0.5.0",
    "jupyter>=1.0.0",
    "ipykernel>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agent", "tools"]

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
target-version = "py310"
